requirements.txt
Flask==3.1.0
flask-cors==5.0.0
gradio_client==2.0.0
python-dotenv==1.0.1
razorpay==1.4.2
google-generativeai==0.8.3
Pillow
numpy
imageio
imageio-ffmpeg
opencv-python
